requirements.txt
ruff==0.14.*
neo4j==5.28.*
python-dotenv==1.1.*
pyzotero==1.6.*
pydantic==2.11.*
polars==1.31.*
httpx==0.28.*
tqdm==4.67.*